+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 Next
 ----
 
+2019.06.03.0
+------------
+
 * Added options to choose the amount of memory given to each VM.
 * Fixed a bug which prevented ``minidcos vagrant`` from working when a VM existed with a space in the name.
 * Fixed a bug which prevented ``minidcos vagrant`` from working in some situations when the ``$HOME`` environment variable is not set.